added #'delete-gpu-function & gpu-function
The gpu-function is equivalent to cl's 'function. To make this work I
have also added a make-load-form for func-key. This means it can exist
as an object in source code, yay!

delete-gpu-function allows you to undefine a gpu-function (as you would
expect)

Now I want to add a restart to allow you to do this easily in those
cases you most hit it