Check to see if fbo attachment sizes match
I have just spent FAR too long debugging an issue that turned out to be
the depth attachment having different dimensions to the other
attachments.

According to the spec:

> If the attachment sizes are not all identical, rendering will
> be limited to the largest area that can fit in all of the
> attachments (an intersection of rectangles having a lower left
> of (0 0) and an upper right of (width height) for each attachment).

This means it's going to be unusual to want to have attachments of
different sizes. So by default we will now throw and error explaining
the situtation if we see the sizes don't match.

You can also override this situation by having
`:matching-dimensions nil` in the args of your #'make-fbo call.