Enable choosing conflicted stage name at expand
def-g-> can throw errors if stage names are overloaded.

Hack in some a thing that invokes the debugger at expand time and let's
you pick the func you want.

We also change gen-recompile-func to make it's stage list from the
validated gpipe-args. This is because we can now change the stage name
using the restart