started exploring the uniform baking function
Bake will allow you to generate a new pipeline with specified uniforms
hardcoded. The primary use case for this is to allow people to turn
partial pipelines into compete ones by specifying the functions.

Right now we are just working on the logic to generate the source code
Varjo will compile. It's a very small modification to the
parsed-gpipe-args->v-translate-args function, so when this works it will
replace the original.