made the def-cepl-context macro
One thing that is definitely missing is how to map a texture target to an
index. However, once that is done this should be quite nice.

Oh wait.. what if multiple singluar targets share a cache? Hmm I'll need
to fix that too. Or I'll have to make the array-buffer-binding-id slots
into and indexed cache.. that should be fine actually given the
target->index fix. We shall see