tidying of context + use unwind-protect everywhere
I found a few places where we hadnt been using unwind-protect, the cost
of this being missing is just too damn high, being in an unknown gl
state is fucking knightmarish when debugging.

Also We are still cleaning up the context related files. It's not done
but meh, we are getting there.