inject-func-key to spec->func-key. details below
For a while now func-keys have had creation-forms (via make-load-form)
however we have still be serializing the keys when using them in
macros.

We now add the func-keys themselves to the generated code but use
spec->func-key for the places where it may be a gpu-func-spec.

spec->func-key is essentially identity if called with a func-key and
creates a func-key for any gpu-func-spec passed in.