multipass compilation working again..but it's ugly
I compile once and collect all external functions, I then add them as
local functions surrounding the main body and compile again. I then get
a shader which uses no externals and so the current multipass code keeps
working.

This is obviously the worst.. We need to get something better. However
lark does work again so that it a good proof that it's holding up.