Make new host api with fallback
We needed to add more things to cepl.host but we also dont want to
break the hosts already out there. To handle this we make a new host
api and then fallback to the old stuff if it isnt being used.

The pattern for the hosts is very simple, subclass the host-api
version you support and populate the slots.