switch def-g-> & g-> to defpipeline-g & pipeline-g
..also `glambda` becomes `lambda-g`

`def-g->`, `g->` & `glambda will keep working under end of december 2017
but `defpipeline-g`, `pipeline-g` & `lambda-g` are the replacements.

The reason for this is that I have realized that when I talk about
`def-g->` I have to say 'def-gee-pipe' this is clumsier and less obvious
than `defpipeline-g`. The rest follows out of consistency