Add with-cepl-context macro. *cepl-context* -> #'cepl-context
By moving from *cepl-context* to #'cepl-context we get to do all kinds
of injection and macro fuckery we can only do with functions.

with-cepl-context lets us set a cepl-context for the lexical
scope. This means CL doesnt have to query a dynamic-var a tonne of
times in a single function. We will use this internally to make sure
our nested macros are querying as little as possible.

Most of this commit is just us replacing *cepl-context*
with (cepl-context) everywhere. The real action is in
cepl-context.lisp