Rewrote the texture/sampler caching
Ok so hopefully this will fix some things.

First issue was that the old caching scheme was wrong :D it didnt
handle texture units properly and only worked well becuase we had been
setting all the time (or had too few programs involved).

First thing I decided was that all binding for the purpose of setting
values would be done in tex-unit 0, all binding for rendering would be
done on 1 and higher.

I also noticed that there was no need for users to query the bound
textures so I could drop certain parts of the caching (there is no
current-texture).

I now only bind the sampler locations once, on init of the
pipeline. This is nice as we then only need to make sure the texture &
sampler are bound to that tex-unit in future (and that is handled by
the pipeline)