Optimize fbo binding
This should bring fbos inline with the other cached objects.

with-fbo-bound still does unbind at the end but it manages the 'is
already bound' check itself so it can avoid unbinding in many cases. We
also provide a bunch of specialized context functions so that it can
avoid checking the binding twice.

With this commit we could expose a 'current-fbo' function in the same
vein as current viewport without any major worries regarding performance
or behaviour.