Add support for the stencil buffer
A note of stencil-params struct implementation.

stencil-params are read only

We only store the enum values in the struct. Set/get/print all
covert the values. This is a bit of an experiment but the idea is
that the most frequent operation to be run on stencil-params will
be to apply them, so that is the usecase we should optimize for.

Due to the above we have made dedicated functions for the enum
translation that will be much faster than the cffi ones

Their read only nature plays well with this as then people are not
encouraged to compare and set fields on existing params. The best
use case is to make one every time you need different stencil
behaviour and apply as needed.