Standardize how we do context-accessors
Now we have a pattern for inlining of the context lookup we want to help
the implementation to use it as much as possible.

with-cepl-context makes a local macro for (cepl-context) which will
expand the a local variable holding the context, this means there is
only one lookup in that local scope.

All context accessor take the context as an optional arg, the default is
to call (cepl-context). However I'm not 100% sure that the optional arg
would be inline in the source and thus be expanded by the local macro.

To help things along we have a compiler macro that explicitly inserts
the call in cases where the context is not provided, this will make it
subject to localmacro expansion and thus we get the boost we are looking
for.