Add release mode. Use release-unwind-protect everywhere in CEPL
This is a first step into the world of having release builds. For now
it just replaces unwind-protect with multiple-value-prog1. Over time
we will add more things to remove safety nets for potential
optimizations.

So far this didnt seem to have any effect so we may remove it
again. We need larger projects to test.