Non-shared multi-context support
Allow up to 8 contexts (WIP)

this limitation is dumb but it makes this first pass of the
implementation easy. Later on we will need to handle any number of
contexts.

We are now going to give each cepl-context an ID between 0 and
+max-context-count+. We will use this ID to index into a per context
cache of program-ids. Because cepl contexts are bound to a single
thread this means that the named pipelines are threadsafe (and uniform
caching is done per context)

As hinted above lambda pipelines are bound to the context in which
they are created

Some Interesting Points:

- we dont support shared contexts yet due to some of the hosts not
  supporting it (glop for example)

- the pipeline-name to id mapping now lives in a global var rather
  than on the context. This is because the name now maps to a lookup
  table (a vector in our case) and this is shared between all
  contexts (and thus threads)

- We track the currently bound program on the cepl-context which means
  our use-program function takes a context.

- Our current use of locks is pretty shitty. We need to do a real
  review and try and make this thing threadsafe (where that doesnt
  impede GL)

- We now show the progam-id when you 'touch' the pipeline