Allow up to 8 contexts (WIP)
this limitation is dumb but it makes this first pass of the
implementation easy. Later on we will need to handle any number of
contexts.

We are now going to give each cepl-context an ID between 0 and
+max-context-count+. We will use this ID to index into a per context
cache of program-ids. Because cepl contexts are bound to a single
thread this means that the named pipelines are threadsafe (and uniform
caching is done per context)

As hinted above lambda pipelines are bound to the context in which
they are created