Make fixing of tfs array error easier from repl
If you are in the scope of a with-transform-feedback you cannot change
which arrays are bound to the transform-feedback-stream. This is good
in general as it prevents bugs but when we have an incorrect number of
arrays bound it is annoying that we can't fix the issue before
'continuing'.

The solution is that if you set the arrays whilst in the scope of
with-transform-feedback we will do the following

- emit a warning explaining the issue
- cache the arrays
- set the array at the end of the transform-feedback-stream

This makes working from the REPL that tiny bit easier