did a little test to evaluate closure -v- defvar access times
I want to move away from the top-level closure but I was worried about
potential slowdown from using a special var.

Luckily it turns out that if we pack the data into a custom struct and
declare the types of the defvar the generated code is faster than
having a top level closure with many vars.

The closure won when there was only 1 var, but in the case we are
interested in it looks like this is going to be fine.

The gains are on the order of microseconds per 5000 draw calls, so
piss all really, but I just could stand the idea of degradation.