Add queries to CEPL
All queries except timestamp are scoped, use with-query-bound to query
everything within the scope and then use #'pull-query-result to get
the value. You can also check if the result is available.

Rather than pull the result back you can push it to a gpu-array, which
is awesome as the data stays on the gpu and you can use it for
dispatching further render tasks