Fix case where redefining uniforms caused state errors
We were getting errors when adding or removing uniforms in some cases.
This change swaps out the defvar/defparameter code in favor of having
the recompile function handling the resetting of the state variable.

I'm not 100% happy with this, it feels data-racy but it seems to be
working well so will be good enough for this release