Rewrite λpipelines to avoid need for implicit uniform uploader func
First aknowledgement is that there is always gonna be a call to
compile if we are using implicit uniforms; either for the
lambda-pipeline or for the uniform-uploader. This means we may as well
make sure it only happens once.

To do this we drop the idea that we will inline-expand gpu-lambdas
when the args are statically known. We now are always going to be
compiling the pipeline body at runtime. The good news is that with a
little shuffling we can remove the need for the 'implicit uniform
uploader'. This is because we can compile the stages and then look at
the actual implicit uniforms required and then emit the code with all
the implicit stuff inlined just like for the regular uniforms. This
then gets compiled with compile.

Of course some of the values that get closed over are from (or are
computed from) the results of compilation, to handle this the emitted
lambda now takes those values as arguments.