Fix silent gpu-func-spec polution bug
The symptom of this bug was that simply expanding a lambda-g form
could do *something* that caused totally seperate lambda-pipelines to
have the wrong code.

The reason for this was twofold:

- lambda-g was inadvertently adding a gpu-func-spec to the cache (with
  name nil) this would then cause issues with any other
  lambda-pipeline with the same args because the spec name would clash
  AND ..

- becuase our gpu-lambda/func-key -> gpu-func-spec code was splattered
  all over the place and so we would end up calling #'gpu-func-spec on
  a gpu-func-spec object, which would then pull the incorrectly cached
  gubbage from above.

part of the solution to this was to do the resolving to gpu-func-specs
as early as possible. This has resulted in a bunch of tweaks to the
internal pipeline code but it should be a less buggy (although it's
still not really any prettier)