Fix occasional recompile pipeline state issues
We were getting into cases where the pipeline function is recompiled
on one thread, and before the thread state is changed the pipeline is
executed. This causes crashes as it tries to index the arrays state
object incorrectly.

To fix thsi we subject ourselves to one more 'if' in the pipeline
object. I'm not too happy with this but for now I'm sure of a better
way. We store a version/diff number in the state object and compare it
with the number inlined in the pipeline, if they dont match then we
need to call init again.

We ditch the state object clearing from the recompile function and
only call eval when there are changes.

All of this should add up to a more stable solution that last time.