Add with-context-state-restored macro
This allows the user to pick certain categories to be reapplied from
the context. This is intended to be used when calling out to other
libraries that manipulate GL state.

It does not set the state to known values before the body but instead
only goes through the context resetting the state requested