First shot add implementing live recompile for lambda pipelines
If this works it will be pretty cool.

The initial issue arose when making a gpu graphing feature. We were
using bake-uniforms to complete pipelines but, as cool as it was, it
did suck that the graph didnt respond to changes to used functions.

I decided that the cost of an `apply` and a struct slot access were
worth it for this feature as long as we could opt out.

As a neat side effect we no longer have to explain the discrepency
between lambda & regular pipelines, they all live recompile now. We
just need to add `:static` to defpipeline-g and explain that the cost for
lambda-pipelines not using `:static:` is non-trivial and we are
sorted!