put type restrictions on functions baked into stage
When bake-uniforms is used we turn the specified args into local
variable that surround the stage code. We have very little validation
and its way to easy to create very confusing errors.

As a first step, when we inject a function, we now surround it with a
`the` form which asserts the type of the function. This at least gives
us a related error.

This came up as I baked a function returned a :vec2 instead of a :vec3
and the compiler gave me errors about some other function not
accepting a :vec2 .. gross.