Empty info of fbos can now be modified and will update gl state
I tried making the empty params a du,,y attachment but it sucked as
that meant it needed a gpu-array-t, so we make a dummy one of
those.. but that needed a dummy texture :D so no lets not go further
down that road of stupid.

Instead we expose the empty-fbo-params and maintain a reference back
to the fbo that owns it. This let's us apply changes to the empty fbo
which in turn allowed a bunch of other cleanups.

I think I'm happy with this for now.