Fix spurious uniform mismatch due to automatic addition of layout
The original issue occured due to the test compile adding :std-140 to
ssbo arguments it didnt have a layout for. This info was being carried
over to the gpu-func-spec.

We call 'add-layout-to-structs' here to ensure that the ubo/ssbo
arguments have the correct layout information. Due to compilation
order we can have calls to aggregate-uniforms at points where the
latest gpu-spec hasnt bee test-compiled.

This means we have to add layout info in a few places rather than
assume the correct info would be available in `defun-g`