Allow enabling rasterized with no fragment stage
This is super handy! It is possible to write to the depth buffer
without a fragment shader (we used this in some gpu occlusion culling
code)

This was not possible previously as CEPL assumed that no fragment
shader meant no rasterizing.

This commit means you can explicitly specify a fragment stage with a
gpu-function designator of nil to keep the rasterizer enabled

In this case there is not fragment stage and the rasterizer is disabled
```
    (defpipeline-g test0 ()
      :vertex (vert :vec2))
```

In this case there is not fragment stage and the rasterizer is enabled
```
    (defpipeline-g test1 ()
      :vertex (vert :vec2)
      :fragment nil)
```

Woo!