fix a number of sbcl notes related to compile order
A bunch of functions are unable to have their types full verified as
functions they depend on have not been compiled by that point in the
file. This reordering and poking addresses that.

A couple of functions were given type definitions where it helped and
didnt change behaviour.

A few cases of unreachable code were also resolved. This was all down
to simple mistakes

The setter for viewport-origin was the biggest change as we now branch
on the incoming value's type to avoid the calls to floor having to
handle generic values. We muffle an unavoidable check though