Add helper functions for populating indirect drawcalls
In this commit we add 3 new functions:

make-draw-command-c-array:

makes a c-array with the element-type of either
`arrays-indirect-command` or `elements-indirect-command` depending on
the nature of the buffer-stream provided

make-draw-command-gpu-array:
same as make-draw-command-c-array but as a gpu-array

set-draw-command:
A function to populate elements of a draw-command array with the
details that are known from the given buffer-stream.

We also made the layout of `arrays-indirect-command` and
`elements-indirect-command` std-430, which lets them be used in
ssbos.

Lastly we exported the symbols that name the slots of the
X-indirect-command structs so they can be used in with-slots in
shaders.