fix stride calculation on types bigger than a vec4
If the data definition via defstruct-g or a built-in type exceeds a vec4 
, the usage of `%gl:vertex-attrib-*pointer` calls changes.

In CEPL, the stride was previously always 0, meaning the generic vertex
attributes are tightly packed (as per the spec). The splitting of a type
into multiple pointers and their offsets was already done. However, in 
both Fedora 37 and macOS 12.6 it gave unexpected results.

The reason, though not documented (AFAICT), is likely due to the stride
being necessary in cases for types longer than a vec4 for vertex 
attributes. By always providing it, we get the expected results.

A sample pipeline is given:
```
(defun-g vert ((data g-pnt)
               (instance :mat4)
               &uniform
               (view :mat4)
               (proj :mat4))
  (* proj view instance (v! (pos data) 1.0)))

(defun-g frag ()
  (vec4 1.0))

(defpipeline-g test-instancing ()
  (vert g-pnt :mat4)
  (frag))
```

The data definition and map-g is left to the patch reader.